[WIP] [skip-ci] Make record to model logic shared
Presentation Record logic shared fully
